Добавлена возможность создавать посты с прикрепленными фотографиями. Добавлена возможность отправить плейлист другу. Исправлено микширование, в больших плейлистах перекключается быстрее. Исправвлена пачка проблем.

+++ VkNet/Model/Attachments/Video.cs
@@ -29,7 +29,7 @@ namespace VkNet.Model.Attachments
 		private long? VideoId
 		{
 			get => Id;
-			set => Id = value;
+			set => Id = (long)value;
 		}
 
 		/// <summary>
